+++ ExilenceNextApp/src/store/priceStore.ts
@@ -1,6 +1,5 @@
 import { AxiosError } from 'axios';
-import { action, computed, makeObservable, observable, toJS } from 'mobx';
-import { persist } from 'mobx-persist';
+import { action, computed, makeObservable, observable } from 'mobx';
 import { fromStream } from 'mobx-utils';
 import { forkJoin, from, interval, of } from 'rxjs';
 import { catchError, concatMap, map, switchMap } from 'rxjs/operators';