+++ ExilenceNextBackend/API/Controllers/StatusController.cs
@@ -1,4 +1,9 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using API.ApiModels;
+using API.Hubs;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.SignalR;
+using Microsoft.Extensions.Configuration;
+using Shared.Entities;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -10,6 +15,13 @@ namespace API.Controllers
     [ApiController]
     public class StatusController : ControllerBase
     {
+        private readonly IHubContext<BaseHub> _hubContext;
+        private readonly IConfiguration _configuration;
+        public StatusController(IConfiguration configuration, IHubContext<BaseHub> hubContext)
+        {
+            _hubContext = hubContext;
+            _configuration = configuration;
+        }
 
         //Used by HAProxy to detect downtime. 
         [Route("")]
@@ -17,6 +29,24 @@ namespace API.Controllers
         public IActionResult GetStatus()
         {
             return Ok();
+        }
+
+        [Route("announcement")]
+        [HttpPost]
+        public async Task<IActionResult> Announcement(AnouncementModel announcement)
+        {
+            var password = _configuration.GetSection("Announcement")["Password"];
+            var message = new AnouncementMessageModel()
+            {
+                Title = announcement.Title,
+                Message = announcement.Message
+            };
+
+            if (password == null || announcement.Password != password)
+                return BadRequest(new { result = "Wrong password" });
+
+            await _hubContext.Clients.All.SendAsync("OnAnnouncement", message);
+            return Ok(new { result = "Message sent" });
         }
     }
 }